Add Pregreeted flag to Client
Set the Pregreeted bool flag on Client to true if the client pregreeted.
A Client object is part of an Envelope, so the user can check if the
Client pregreeted on delivery.

Perhaps the other callbacks (OnMailFrom, etc.) should be able to check
this flag too. They would need to be passed the Client object somehow.
Maybe a Pregreeted() bool func could be added to the Connection
interface.

Add tests for pregreet checking. TestPregreet seems to hang inexplicably
once in a while.